refactor : 각 모듈별 스프링 통합테스트를 위한 커스텀 어노테이션 생성
