+++ Whatnow-Api/src/test/kotlin/com/depromeet/whatnow/WhatnowApiApplicationTests.kt
@@ -1,0 +1,12 @@
+package com.depromeet.whatnow
+
+import com.depromeet.whatnow.config.ApiIntegrateSpringBootTest
+import org.junit.jupiter.api.Test
+
+@ApiIntegrateSpringBootTest
+class WhatnowApiApplicationTests {
+
+    @Test
+    fun contextLoads() {
+    }
+}
