+++ Whatnow-Api/src/test/kotlin/com/depromeet/whatnow/config/ApiIntergrateProfileResovler.kt
@@ -1,0 +1,10 @@
+package com.depromeet.whatnow.config
+
+import org.springframework.test.context.ActiveProfilesResolver
+
+class ApiIntegrateProfileResolver : ActiveProfilesResolver {
+
+    override fun resolve(testClass: Class<*>): Array<String> {
+        return arrayOf("infrastructure", "domain")
+    }
+}
