--- Whatnow-Api/src/test/kotlin/com/depromeet/whatnowapi/PromiseApiApplicationTests.kt
@@ -1,12 +1,0 @@
-package com.depromeet.whatnowapi
-
-import org.junit.jupiter.api.Test
-import org.springframework.boot.test.context.SpringBootTest
-
-@SpringBootTest
-class PromiseApiApplicationTests {
-
-    @Test
-    fun contextLoads() {
-    }
-}
