+++ Whatnow-Domain/src/test/kotlin/com/depromeet/whatnow/WhatnowDomainApplicationTests.kt
@@ -1,9 +1,10 @@
-package com.depromeet.whatnowdomain
+package com.depromeet.whatnow
 
+import com.depromeet.whatnow.config.DomainIntegrateSpringBootTest
 import org.junit.jupiter.api.Test
 import org.springframework.boot.test.context.SpringBootTest
 
-@SpringBootTest
+@DomainIntegrateSpringBootTest
 class WhatnowDomainApplicationTests {
 
     @Test
