+++ Whatnow-Domain/src/test/kotlin/com/depromeet/whatnow/config/DomainIntegrateTestConfig.kt
@@ -1,0 +1,11 @@
+package com.depromeet.whatnow.config
+
+import com.depromeet.whatnow.WhatnowDomainApplication
+import com.depromeet.whatnow.WhatnowInfrastructureApplication
+import org.springframework.context.annotation.ComponentScan
+import org.springframework.context.annotation.Configuration
+
+@Configuration
+@ComponentScan(basePackageClasses = [WhatnowInfrastructureApplication::class, WhatnowDomainApplication::class])
+class DomainIntegrateTestConfig {
+}
