+++ Whatnow-Domain/src/test/kotlin/com/depromeet/whatnow/config/DomainIntergrateProfileResovler.kt
@@ -1,0 +1,10 @@
+package com.depromeet.whatnow.config
+
+import org.springframework.test.context.ActiveProfilesResolver
+
+class DomainIntegrateProfileResolver : ActiveProfilesResolver {
+
+    override fun resolve(testClass: Class<*>): Array<String> {
+        return arrayOf("infrastructure", "domain")
+    }
+}
