+++ Whatnow-Domain/src/test/kotlin/com/depromeet/whatnow/domains/example/ExampleClassTest.kt
@@ -1,6 +1,5 @@
-package com.depromeet.whatnowdomain.domains.example
+package com.depromeet.whatnow.domains.example
 
-import com.depromeet.whatnow.domains.example.ExampleClass
 import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.Test
