+++ Whatnow-Infrastructure/src/test/kotlin/com/depromeet/whatnow/WhatnowInfrastructureApplicationTests.kt
@@ -1,4 +1,4 @@
-package com.depromeet.whatnowinfrastructure
+package com.depromeet.whatnow
 
 import org.junit.jupiter.api.Test
 import org.springframework.boot.test.context.SpringBootTest
